+++ package.json
@@ -13,8 +13,8 @@
     "lint": "prettier --check . && eslint scripts/ tests/",
     "gen-csv": "tsx scripts/generateCsv.ts",
     "gen-html": "tsx scripts/generateHtmlPages.ts",
+    "sync-core-data": "tsx scripts/syncCoreData.ts",
     "sync-extended-data": "tsx scripts/syncExtendedData.ts",
-    "update-map-data": "tsx scripts/updateMapData.ts",
     "sync-lat-lng": "tsx scripts/syncLatLng.ts",
     "broken-links": "tsx scripts/brokenLinks.ts",
     "serve-dist": "cd dist; http-server",
