+++ scripts/syncCoreData.ts
